src/discord-ai-bot-trivia-greetings/requirements.txt
discord.py==2.4.0
openai==1.61.1
python-dotenv==1.0.1
